reorganize project structure, add accraual system loop
