+++ cmd/gophermart/main.go
@@ -1,18 +1,15 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
-	"time"
 
+	"github.com/adettelle/loyalty-system/internal/accrualservice"
 	"github.com/adettelle/loyalty-system/internal/gophermart/api"
 	"github.com/adettelle/loyalty-system/internal/gophermart/config"
 	"github.com/adettelle/loyalty-system/internal/gophermart/database"
-	"github.com/adettelle/loyalty-system/internal/gophermart/model"
 )
 
 func main() {
@@ -22,7 +19,6 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.Println(config)
 
 	if config.DBUri != "" {
 		uri = config.DBUri
@@ -51,92 +47,45 @@ func main() {
 
 	r := api.NewRouter(storage)
 
-	go func() {
-		ticker := time.NewTicker(time.Second * 2)
+	accrualSystem := accrualservice.NewAccrualSystem(db, config.AccrualSystemAddress)
 
-		for range ticker.C {
-			ordersWithNewStatus, err := model.GetAllNewOrders(db, context.Background())
-			log.Println("Orders with new status:", ordersWithNewStatus)
-			if err != nil {
-				log.Println("err1:", err) // ??????????????????????
-				continue
-			}
-			for _, ord := range ordersWithNewStatus {
-				orderFromAccrual, err := GetOrderFromAccrualSystem(ord.Number, config.AccrualSystemAddress)
-				if err != nil {
-					log.Println("err2:", err) // ????????????????????
-					continue
-				}
-				log.Println("Orders from accrual system:", orderFromAccrual)
+	accrualSystem.AccrualLoop()
 
-				err = model.UpdateOrderStatus(orderFromAccrual.Status, ord.Number, db, context.Background())
-				if err != nil {
-					log.Println("err3:", err) // ????????????????????
-					continue
-				}
+	// go func() {
+	// 	ticker := time.NewTicker(time.Second * 2)
 
-				err = model.UpdateAccrualPoints(orderFromAccrual.Accrual, ord.Number, db, context.Background())
-				if err != nil {
-					log.Println("err4:", err) // ????????????????????
-					continue
-				}
-			}
-		}
-	}()
+	// 	for range ticker.C {
+	// 		ordersWithNewStatus, err := model.GetAllNewOrders(db, context.Background())
+	// 		log.Println("Orders with new status:", ordersWithNewStatus)
+	// 		if err != nil {
+	// 			log.Println("err1:", err) // ??????????????????????
+	// 			continue
+	// 		}
+	// 		for _, ord := range ordersWithNewStatus {
+	// 			orderFromAccrual, err := accrualservice.GetOrderFromAccrualSystem(ord.Number, config.AccrualSystemAddress)
+	// 			if err != nil {
+	// 				log.Println("err2:", err) // ????????????????????
+	// 				continue
+	// 			}
+	// 			log.Println("Orders from accrual system:", orderFromAccrual)
+
+	// 			err = model.UpdateOrderStatus(orderFromAccrual.Status, ord.Number, db, context.Background())
+	// 			if err != nil {
+	// 				log.Println("err3:", err) // ????????????????????
+	// 				continue
+	// 			}
+
+	// 			err = model.UpdateAccrualPoints(orderFromAccrual.Accrual, ord.Number, db, context.Background())
+	// 			if err != nil {
+	// 				log.Println("err4:", err) // ????????????????????
+	// 				continue
+	// 			}
+	// 		}
+	// 	}
+	// }()
 
 	err = http.ListenAndServe(address, r)
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-type OrderStatsResp struct {
-	Order   string  `json:"order"`
-	Status  string  `json:"status"`
-	Accrual float64 `json:"accrual,omitempty"`
-}
-
-// GET /api/orders/{number}
-func GetOrderFromAccrualSystem(number string, url string) (OrderStatsResp, error) {
-	var ord OrderStatsResp
-
-	url = url + "/api/orders/" + number
-	log.Println(url)
-	req, err := http.NewRequest(http.MethodGet, url, nil)
-	if err != nil {
-		log.Println("err1:", err)
-		return OrderStatsResp{}, err
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		log.Println("err2:", err)
-		return OrderStatsResp{}, err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return OrderStatsResp{}, fmt.Errorf("error getting order from accrual system: order %s, ststus %d", number, resp.StatusCode)
-	}
-
-	var buf bytes.Buffer
-
-	_, err = buf.ReadFrom(resp.Body)
-	if err != nil {
-		log.Println("err3:", err)
-		return OrderStatsResp{}, err
-	}
-
-	if err := json.Unmarshal(buf.Bytes(), &ord); err != nil {
-		log.Println("err4:", err)
-		return OrderStatsResp{}, err
-	}
-
-	resp.Body.Close()
-
-	log.Println(resp.StatusCode)
-	log.Println("ord:", ord)
-
-	resp.Header.Set("Content-Type", "application/json")
-
-	return ord, nil
 }
